chore: moved helper files (helper, parser, recommender) into helpers folder
