+++ packages/formula-editor/src/formula-editor.ts
@@ -1,7 +1,7 @@
 import { html, LitElement, PropertyValueMap, PropertyValues } from "lit";
 import { customElement, property, state, query } from "lit/decorators.js";
 import { FormulaEditorStyles } from "./styles/editor.js";
-import { Parser } from "./parser.js";
+import { Parser } from "./helpers/parser.js";
 import "./suggestion-menu.js";
 
 @customElement("formula-editor")
