+++ packages/formula-editor/src/helpers/helpers.ts
