+++ packages/formula-editor/src/helpers/parser.ts
